create alternative profileSlice, move firestoreAPI to services folder
