+++ src/Registration/ProfileForm.tsx
@@ -8,9 +8,9 @@ import { Guser, Profile } from "../app/model";
 import { LocalizationProvider } from "@mui/x-date-pickers";
 import { AdapterDayjs } from "@mui/x-date-pickers/AdapterDayjs";
 import { useDispatch, useSelector } from "react-redux";
-import { selectProfile, addProfile } from "../features/ProfileSlice";
+import { selectProfile, addProfile } from "../features/profileSlice";
 import { Button } from "@mui/material";
-import { GetDataByTwoColumns2, PostData, GetDataByTwoColumns } from "../app/functions/DbFunctions";
+import { GetDataByTwoColumns2, PostData, GetDataByTwoColumns } from "../app/services/DbFunctions";
 import { ToastContainer, toast, ToastOptions } from "react-toastify";
 import 'react-toastify/dist/ReactToastify.css';
 import { selectUser } from "../features/userSlice";
